+++ animation_nodes/ui/node_menu.py
@@ -367,7 +367,6 @@ class MeshMenu(bpy.types.Menu):
         insertNode(layout, "an_MeshInfoNode", "Mesh Info")
         insertNode(layout, "an_ObjectBoundingBoxNode", "Get Bounding Box")
         insertNode(layout, "an_CombineMeshNode", "Combine Mesh")
-        insertNode(layout, "an_VertexGroupInputNode", "Vertex Group Input")
         insertNode(layout, "an_MeshFromSplineNode", "Mesh From Spline")
         layout.separator()
         insertNode(layout, "an_VertexGroupInputNode", "Vertex Group Input")